+++ core/validators/topology_snapshot_test.go
@@ -15,7 +15,6 @@ package validators_test
 import (
 	"bytes"
 	"context"
-	"encoding/hex"
 	"testing"
 
 	"code.vegaprotocol.io/vega/core/validators"
@@ -79,7 +78,7 @@ func TestTopologySnapshot(t *testing.T) {
 	nr1 := commandspb.AnnounceNode{
 		Id:              "vega-master-pubkey",
 		ChainPubKey:     tmPubKeys[0],
-		VegaPubKey:      hex.EncodeToString([]byte("vega-key")),
+		VegaPubKey:      hexEncode("vega-key"),
 		EthereumAddress: "0x6d53C489bbda35B8096C8b4Cb362e2889F82E19B",
 	}
 	err = top.AddNewNode(ctx, &nr1, validators.ValidatorStatusTendermint)
@@ -88,7 +87,7 @@ func TestTopologySnapshot(t *testing.T) {
 	nr2 := commandspb.AnnounceNode{
 		Id:              "vega-master-pubkey-2",
 		ChainPubKey:     tmPubKeys[1],
-		VegaPubKey:      hex.EncodeToString([]byte("vega-key-2")),
+		VegaPubKey:      hexEncode("vega-key-2"),
 		EthereumAddress: "0x6d53C489bbda35B8096C8b4Cb362e2889F82E19B",
 	}
 	err = top.AddNewNode(ctx, &nr2, validators.ValidatorStatusTendermint)
@@ -117,7 +116,7 @@ func TestTopologySnapshot(t *testing.T) {
 		CurrentAddress: "0x6d53C489bbda35B8096C8b4Cb362e2889F82E19B",
 		NewAddress:     "0x69bA3B3e6B5b1226A2e26De9a9E2D9C98f2b144B",
 	}
-	err = top.RotateEthereumKey(ctx, nr1.Id, 5, ekr1)
+	err = top.RotateEthereumKey(ctx, hexEncode("vega-key"), 5, ekr1)
 	assert.NoError(t, err)
 
 	ekr2 := &commandspb.EthereumKeyRotateSubmission{
@@ -125,7 +124,7 @@ func TestTopologySnapshot(t *testing.T) {
 		CurrentAddress: "0x6d53C489bbda35B8096C8b4Cb362e2889F82E19B",
 		NewAddress:     "0xd6B6e9514f2793Af89745Fd69FDa0DAbC228d336",
 	}
-	err = top.RotateEthereumKey(ctx, nr2.Id, 5, ekr2)
+	err = top.RotateEthereumKey(ctx, hexEncode("vega-key-2"), 5, ekr2)
 	assert.NoError(t, err)
 
 	// Check the hashes have changed after each state change